Added post-stimuli blanking window
Added a post-stimulation blanking window that can be set by the user under the highpass filter setting. It keeps the highpass filter output to 0 after an electrical stimulation for the duration choose by the user in millisecond. This acts on both the displayed high-pass trace and on the spike detection.